audio_processing/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
python-dotenv==1.0.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
pydub>=0.25.1
ffmpeg-python>=0.2.0
scipy>=1.11.0
discord.py>=2.3.0